Reworked escapes. Made entry separator double as end of value character.
Added newline, tab, etc. escapes, Unicode escapes, disallowed all escapes that are not predefined.
Entry separator now also works as an "end of value" character. This is useful if you want to parse a Qson value from a larger string: just append a ' to signal that the Qson value ends there.